Refactor : refactor codes (2)
